feat(skeleton-row): add component for table loading state (#435)
* feat(tablebody): create SkeletonTable and implement it in TableBody

Implemented loading prop in TableBody to render Skeleton rows while async data is being loaded.
For that, it was necessary to create to new Skeleton components: SkeletnTable and SkeletonCellre #433

* refactor(skeletoncontainer): update id of the animate tag to be unique

the SkeletonTable renders several skeletons and with that the id of the animate tag repeats. Since
id should be unique in HTML tags the browser and the tests will complain about it

* test(skeleton-table): add snapshot test for skeleton table

* chore(size-limit): increase size-limit to 305KB and update snapshots

* refactor(skeleton-table): change prop names and add knobs to story

Change prop names to names that follow our conventions. Also implemented knobs in their stories

* refactor(skeleton-row): rename skeleton table to skeleton row

* refactor(table): make skeleton row into functional component

Refactor the skeleton table components into functional components and remove unecessary code

* test(table): upload table related snapshots

* refactor(table): simplify loading state for tables

Reverted changes in TableBody. This way, the loading state is controlled by two props in the SkeletonRow component. This makes the API easier to work with

* test(skeleton-row): update snapshots