fix(dropdown): use capture event in outside click handler
This fix handles the situation where the dropdown is open and the user then clicks on the toggle element to close the dropdown, which fires both an open and a close event right after each other.

This happens as the event bubbles from the button and up to the document, where the outside click event handler is bound. The alternative is using a capture event like in this PR, where the event goes from the document and down to the button. This solves the issue of two events firing, but the downside is that it now also takes a second click to e.g. open another dropdown; the first to close the active dropdown and the second to trigger the event of the action.