feat(component): added start and end icon to Badge
Icons used before and after the badge content, but still within the Badge itself. The padding and
position had to be adjusted manually to allow the icons to fit neatly with center alignment along
the horizontal axis. The Icon sizes used are small for the default sized badge and tiny for the
small sized badge. Since the tiny sized icons are the same height as the small sized badge (16px),
some of the icons barely fit.

"#2125"

Committer: Wake1st <joeltwakefield@gmail.com>